feat: pagination css var
