+++ components/pagination/Pagination.tsx
@@ -14,6 +14,8 @@ import { useLocale } from '../locale';
 import { MiddleSelect, MiniSelect } from './Select';
 import useStyle from './style';
 import useCSSVar from './style/cssVar';
+import { useToken } from '../theme/internal';
+import BorderedStyle from './style/bordered';
 
 export interface PaginationProps extends RcPaginationProps {
   showQuickJumper?: boolean | { goButton?: React.ReactNode };
@@ -50,6 +52,7 @@ const Pagination: React.FC<PaginationProps> = (props) => {
     ...restProps
   } = props;
   const { xs } = useBreakpoint(responsive);
+  const [, token] = useToken();
 
   const { getPrefixCls, direction, pagination = {} } = React.useContext(ConfigContext);
   const prefixCls = getPrefixCls('pagination', customizePrefixCls);
@@ -113,6 +116,7 @@ const Pagination: React.FC<PaginationProps> = (props) => {
     {
       [`${prefixCls}-mini`]: isSmall,
       [`${prefixCls}-rtl`]: direction === 'rtl',
+      [`${prefixCls}-bordered`]: token.wireframe,
     },
     pagination?.className,
     className,
@@ -123,17 +127,20 @@ const Pagination: React.FC<PaginationProps> = (props) => {
   const mergedStyle: React.CSSProperties = { ...pagination?.style, ...style };
 
   return wrapCSSVar(
-    <RcPagination
-      {...iconsProps}
-      {...restProps}
-      style={mergedStyle}
-      prefixCls={prefixCls}
-      selectPrefixCls={selectPrefixCls}
-      className={extendedClassName}
-      selectComponentClass={selectComponentClass || (isSmall ? MiniSelect : MiddleSelect)}
-      locale={locale}
-      showSizeChanger={mergedShowSizeChanger}
-    />,
+    <>
+      {token.wireframe && <BorderedStyle prefixCls={prefixCls} />}
+      <RcPagination
+        {...iconsProps}
+        {...restProps}
+        style={mergedStyle}
+        prefixCls={prefixCls}
+        selectPrefixCls={selectPrefixCls}
+        className={extendedClassName}
+        selectComponentClass={selectComponentClass || (isSmall ? MiniSelect : MiddleSelect)}
+        locale={locale}
+        showSizeChanger={mergedShowSizeChanger}
+      />
+    </>,
   );
 };
